requirements.txt
fastapi
uvicorn[standard]
weaviate-client
python-dotenv
requests
